LayoutTests/imported/w3c/web-platform-tests/fetch/corb/style-css-with-json-parser-breaker.sub-expected.txt

Header example

Paragraph body


FAIL Untitled Error: assert_equals: expected "rgb(255, 0, 0)" but got "rgb(0, 128, 0)"
